This enables somewhat hacky SVG support.
A full solution would include first-class support for namespaces throughout the
DOM. But there are only three SVG elements that clash with HTML elements (a,
script, title) so this commit simply assumes that all the other SVG elements
should be in the SVG namespace, and everything else defaults as before.

In addition, we now need to track these nodes, since SVG class names are
different (they are instances of SVGAnimatedString) so the set_class_name
optimization needs to be disabled for these. Tracking this unfortunately
imposes some extra cost on the DOM representation.